Use most up to date value from storage
