+++ src/features/customerList/customerListTableFilters/CustomerListTableFilters.tsx
@@ -1,7 +1,6 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { IconAngleDown } from 'hds-react';
 import { useTranslation } from 'react-i18next';
-import { useHistory, useLocation } from 'react-router-dom';
 
 import useListTableFilters from './useListTableFilters';
 import CustomerListTableFiltersFormContainer from './CustomerListTableFiltersFormContainer';
@@ -15,7 +14,6 @@ enum Content {
 }
 
 const CustomerListTableFilters = () => {
-  usePersistedSearch();
   const { t } = useTranslation();
   const [areFiltersExpanded, setFiltersExpanded] = React.useState<boolean>(false);
   const [listTableFilters] = useListTableFilters();
@@ -53,27 +51,6 @@ function getVariant(areFiltersExpanded: boolean, someFilterIsActive: boolean) {
   }
 
   return Content.EMPTY;
-}
-
-const PERSISTED_SEARCH_SESSION_STORAGE_KEY = 'berth-reservations-admin/persistedCustomerListTableFilters';
-
-function usePersistedSearch() {
-  const { search, pathname } = useLocation();
-  const history = useHistory();
-  const persistedSearch = sessionStorage.getItem(PERSISTED_SEARCH_SESSION_STORAGE_KEY);
-
-  useEffect(() => {
-    // Whenever search changes, persist it for the duration of the session
-    sessionStorage.setItem(PERSISTED_SEARCH_SESSION_STORAGE_KEY, search);
-  }, [search]);
-
-  useEffect(() => {
-    // If search is empty and we have a persisted search, apply the persisted
-    // search
-    if (!search && persistedSearch) {
-      history.replace(`${pathname}${persistedSearch}`);
-    }
-  }, [history, pathname, persistedSearch, search]);
 }
 
 export default CustomerListTableFilters;
